+++ src/interfaces/PokemonApiInterface.ts
